+++ src/Repository/Evaluation/ResourceRepository.php
@@ -1,0 +1,21 @@
+<?php
+
+namespace App\Repository\Evaluation;
+
+use App\Entity\Evaluation\Resource;
+use Doctrine\Bundle\DoctrineBundle\Repository\ServiceEntityRepository;
+use Doctrine\Persistence\ManagerRegistry;
+
+/**
+ * @method Resource|null find($id, $lockMode = null, $lockVersion = null)
+ * @method Resource|null findOneBy(array $criteria, array $orderBy = null)
+ * @method Resource[]    findAll()
+ * @method Resource[]    findBy(array $criteria, array $orderBy = null, $limit = null, $offset = null)
+ */
+class ResourceRepository extends ServiceEntityRepository
+{
+    public function __construct(ManagerRegistry $registry)
+    {
+        parent::__construct($registry, Resource::class);
+    }
+}